+++ src/main/java/com/nusiss/userservice/dao/AddressRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface AddressRepository extends JpaRepository<Address, Long> {
+public interface AddressRepository extends JpaRepository<Address, Integer> {
 
-    List<Address> findByUserId(Long userId);
+    List<Address> findByUserId(Integer userId);
 }
